some changes to fix compilation issues on cygwin: broke up kws-functions.cc into two pieces; add flag in makefile.  Plus a cosmetic change in nnet-computation-graph.cc.
