+++ src/kws/kws-functions.h
@@ -51,37 +51,37 @@ class Interval {
 // We define a function bool CompareInterval(const Interval &i1, const Interval
 // &i2) to compare the Interval defined above. If interval i1 is in front of
 // interval i2, then return true; otherwise return false.
-bool CompareInterval(const Interval &i1, 
-                     const Interval &i2); 
+bool CompareInterval(const Interval &i1,
+                     const Interval &i2);
 
 // This function clusters the arcs with same word id and overlapping time-spans.
 // Examples of clusters:
-// 0 1 a a (0.1s ~ 0.5s) and 2 3 a a (0.2s ~ 0.4s) are within the same cluster; 
-// 0 1 a a (0.1s ~ 0.5s) and 5 6 b b (0.2s ~ 0.4s) are in different clusters; 
+// 0 1 a a (0.1s ~ 0.5s) and 2 3 a a (0.2s ~ 0.4s) are within the same cluster;
+// 0 1 a a (0.1s ~ 0.5s) and 5 6 b b (0.2s ~ 0.4s) are in different clusters;
 // 0 1 a a (0.1s ~ 0.5s) and 7 8 a a (0.9s ~ 1.4s) are also in different clusters.
 // It puts disambiguating symbols in the olabels, leaving the words on the
 // ilabels.
-bool ClusterLattice(CompactLattice *clat, 
+bool ClusterLattice(CompactLattice *clat,
                     const vector<int32> &state_times);
 
 // This function contains two steps: weight pushing and factor generation. The
 // original ShortestDistance() is not very efficient, so we do the weight
 // pushing and shortest path manually by computing the alphas and betas. The
 // factor generation step expand the lattice to the LXTXT' semiring, with
-// additional start state and end state (and corresponding arcs) added. 
+// additional start state and end state (and corresponding arcs) added.
 bool CreateFactorTransducer(const CompactLattice &clat,
-                            const vector<int32> &state_times, 
-                            int32 utterance_id, 
+                            const vector<int32> &state_times,
+                            int32 utterance_id,
                             KwsProductFst *factor_transducer);
 
 // This function removes the arcs with long silence. By "long" we mean arcs with
 // #frames exceeding the given max_silence_frames. We do this filtering because
-// the gap between adjacent words in a keyword must be <= 0.5 second. 
+// the gap between adjacent words in a keyword must be <= 0.5 second.
 // Note that we should not remove the arcs created in the factor generation
 // step, so the "search area" is limited to the original arcs before factor
-// generation. 
-void RemoveLongSilences(int32 max_silence_frames, 
-                        const vector<int32> &state_times, 
+// generation.
+void RemoveLongSilences(int32 max_silence_frames,
+                        const vector<int32> &state_times,
                         KwsProductFst *factor_transducer);
 
 // Do the factor merging part: encode input and output, and apply weighted
@@ -107,6 +107,40 @@ void OptimizeFactorTransducer(KwsLexicographicFst *index_transducer,
 // GetVerboseLevel() >= 2 and a substantially negative cost is found.
 void MaybeDoSanityCheck(const KwsProductFst &factor_transducer);
 void MaybeDoSanityCheck(const KwsLexicographicFst &index_transducer);
+
+
+// this Mapper class is used in some of the the internals; we have to declare it
+// in the header because, for the sake of compilation time, we split up the
+// implementation into two .cc files.
+class KwsProductFstToKwsLexicographicFstMapper {
+ public:
+  typedef KwsProductArc FromArc;
+  typedef KwsProductWeight FromWeight;
+  typedef KwsLexicographicArc ToArc;
+  typedef KwsLexicographicWeight ToWeight;
+
+  KwsProductFstToKwsLexicographicFstMapper() {}
+
+  inline ToArc operator()(const FromArc &arc) const {
+    return ToArc(arc.ilabel,
+                 arc.olabel,
+                 (arc.weight == FromWeight::Zero() ?
+                  ToWeight::Zero() :
+                  ToWeight(arc.weight.Value1().Value(),
+                           StdLStdWeight(arc.weight.Value2().Value1().Value(),
+                                         arc.weight.Value2().Value2().Value()))),
+                 arc.nextstate);
+  }
+
+  fst::MapFinalAction FinalAction() const { return fst::MAP_NO_SUPERFINAL; }
+
+  fst::MapSymbolsAction InputSymbolsAction() const { return fst::MAP_COPY_SYMBOLS; }
+
+  fst::MapSymbolsAction OutputSymbolsAction() const { return fst::MAP_COPY_SYMBOLS;}
+
+  uint64 Properties(uint64 props) const { return props; }
+};
+
 
 
 } // namespace kaldi
